Move statik pkg into the pkg/ folder
